Add merlin + base role class
